+++ editor/AddSamplesPrompt.ts
@@ -1,11 +1,11 @@
 import { HTML, SVG } from "imperative-html/dist/esm/elements-strict";
 import { Dictionary, Config } from "../synth/SynthConfig";
-import { clamp, parseFloatWithDefault, parseIntWithDefault } from "../synth/synth";
+import { clamp } from "../synth/synth";
 import { ColorConfig } from "./ColorConfig";
 import { EditorConfig } from "./EditorConfig";
 import { SongDocument } from "./SongDocument";
 
-const { div, input, button, a, code, textarea, details, summary, span, ul, li, select, option, h2 } = HTML;
+const { div, input, button, a, code, textarea, details, summary, span, ul, li, h2 } = HTML;
 
 interface SampleEntry {
     url: string;
@@ -15,8 +15,6 @@ interface SampleEntry {
     chipWaveLoopStart: number | null;
     chipWaveLoopEnd: number | null;
     chipWaveStartOffset: number | null;
-    chipWaveLoopMode: number | null;
-    chipWavePlayBackwards: boolean;
 }
 
 interface ParsedEntries {
@@ -25,10 +23,6 @@ interface ParsedEntries {
 
 // @TODO:
 // - Check for duplicate sample URLs and names.
-// - Maybe the Backwards checkbox should be a select as well? Right now though,
-//   assuming that false is the same as if it wasn't actually set should work
-//   fine.
-// - Use constants or an enum for the key-value pairs.
 
 export class AddSamplesPrompt {
     private readonly _maxSamples: number = 64;
@@ -65,7 +59,7 @@ export class AddSamplesPrompt {
     );
     private readonly _closeInstructionsButton: HTMLButtonElement = button({ style: "height: auto; min-height: var(--button-size); width: 100%;" }, "Close instructions");
     private readonly _instructionsArea: HTMLDivElement = div(
-        { style: "display: none; margin-top: 0; -webkit-user-select: text; -moz-user-select: text; -ms-user-select: text; user-select: text; cursor: text; overflow-y: auto;" },
+        { style: "display: none; margin-top: 0; -webkit-user-select: text; -moz-user-select: text; -ms-user-select: text; user-select: text; cursor: text;" },
         h2("Add Samples"),
         div({ style: "margin-top: 0.5em; margin-bottom: 0.5em;" },
             "In UltraBox, custom samples are loaded from arbitrary URLs.",
@@ -91,9 +85,8 @@ export class AddSamplesPrompt {
             "As for where to upload your samples, here are some suggestions:",
             ul({ style: "text-align: left;" },
                 li(a({ href: "https://filegarden.com" }, "File Garden")),
-                li(a({ href: "https://catbox.moe/" }, "Catbox")),
                 li(a({ href: "https://www.dropbox.com" }, "Dropbox"), " (domain needs to be ", code("https://dl.dropboxusercontent.com"), ")"),
-                //li(a({ href: "https://discord.com" }, "Discord"), " (domain needs to be ", code("https://media.discordapp.net"), ")")
+                li(a({ href: "https://discord.com" }, "Discord"), " (domain needs to be ", code("https://media.discordapp.net"), ")")
             )
         ),
         div({ style: "margin-top: 0.5em; margin-bottom: 0.5em;" },
@@ -102,16 +95,16 @@ export class AddSamplesPrompt {
         ),
         div({ style: "margin-top: 0.5em; margin-bottom: 1em;" },
             "Finally, if have a soundfont you'd like to get samples from, consider using this ",
-            a({ href: "./sample_extractor.html", target: "_blank" }, "sample extractor"),
+            a({ href: "https://ultraabox.github.io/sample_extractor.html", target: "_blank" }, "sample extractor"),
             "."
         ),
         div({ style: "display: flex; flex-direction: row-reverse; justify-content: space-between; margin-top: 0.5em;" }, this._closeInstructionsButton)
     );
-    private readonly _addSamplesArea: HTMLDivElement = div({ style: "overflow-y: auto;" },
+    private readonly _addSamplesArea: HTMLDivElement = div(
         h2("Add Samples"),
         div({ style: "display: flex; flex-direction: column; align-items: center; margin-bottom: 0.5em;" },
             this._description,
-            div({ style: "width: 100%; max-height: 450px; overflow-y: scroll;" }, this._entryContainer),
+            div({ style: "width: 100%; max-height: 350px; overflow-y: scroll;" }, this._entryContainer),
             this._addSamplesAreaBottom
         ),
         div({ style: "display: flex; flex-direction: row-reverse; justify-content: space-between;" }, this._okayButton)
@@ -121,7 +114,7 @@ export class AddSamplesPrompt {
     });
     private readonly _bulkAddConfirmButton: HTMLButtonElement = button({ style: "height: auto; min-height: var(--button-size); width: 100%;" }, "Add");
     private readonly _bulkAddArea: HTMLDivElement = div(
-        { style: "display: none; overflow-y: auto;" },
+        { style: "display: none;" },
         h2({ style: "margin-bottom: 0.5em;" }, "Add Multiple Samples"),
         div({ style: "display: flex; flex-direction: column; align-items: center;" },
             div(`Add one URL per line. Remember that you can only have ${this._maxSamples} samples!`),
@@ -130,7 +123,7 @@ export class AddSamplesPrompt {
         ),
         div({ style: "display: flex; flex-direction: row-reverse; justify-content: space-between;" }, this._bulkAddConfirmButton),
     );
-    public container: HTMLDivElement = div({ class: "prompt noSelection", style: "width: 450px; max-height: calc(100% - 100px);" },
+    public container: HTMLDivElement = div({ class: "prompt noSelection", style: "width: 350px;" },
         this._addSamplesArea,
         this._bulkAddArea,
         this._instructionsArea,
@@ -140,10 +133,7 @@ export class AddSamplesPrompt {
     constructor(_doc: SongDocument) {
         this._doc = _doc;
         if (EditorConfig.customSamples != null) {
-            // In this case, `EditorConfig.customSamples` should have the URLs
-            // with the new syntax, as we only use that when saving, and we
-            // also force the new syntax on load, in `Song._parseAndConfigureCustomSample`.
-            const parsed = this._parseURLs(EditorConfig.customSamples, false);
+            const parsed = this._parseURLs(EditorConfig.customSamples);
             this._entries = parsed.entries;
         }
         this._addSampleButton.addEventListener("click", this._whenAddSampleClicked);
@@ -179,7 +169,7 @@ export class AddSamplesPrompt {
     private _saveChanges = (): void => {
         const urlData: string = this._generateURLData();
         EditorConfig.customSamples = urlData.split("|").filter(x => x !== "");
-        Config.willReloadForCustomSamples = true;
+        EditorConfig.willReloadForCustomSamples = true;
         window.location.hash = this._doc.song.toBase64String();
         // The prompt seems to get stuck if reloading is done too quickly.
         setTimeout(() => { location.reload(); }, 50);
@@ -195,8 +185,6 @@ export class AddSamplesPrompt {
             chipWaveLoopStart: null,
             chipWaveLoopEnd: null,
             chipWaveStartOffset: null,
-            chipWaveLoopMode: null,
-            chipWavePlayBackwards: false,
         });
         this._entryOptionsDisplayStates[entryIndex] = false;
         this._reconfigureAddSampleButton();
@@ -223,15 +211,12 @@ export class AddSamplesPrompt {
     private _whenBulkAddConfirmClicked = (event: Event): void => {
         this._addSamplesArea.style.display = "";
         this._bulkAddArea.style.display = "none";
-        // In this case, we shouldn't really bother supporting the old syntax,
-        // as people are only really sharing URLs with the new one.
         const parsed: ParsedEntries = this._parseURLs(
             this._bulkAddTextarea.value
                 .replace(/\n/g, "|")
                 .split("|")
-                .map((x: string) => decodeURIComponent(x.trim()))
-                .filter((x: string) => x !== ""),
-            false
+                // .map((x: string) => decodeURIComponent(x.trim()))
+                .filter((x: string) => x !== "")
         );
         const seen: Map<string, boolean> = new Map();
         for (const entry of this._entries) {
@@ -274,15 +259,21 @@ export class AddSamplesPrompt {
     private _whenSampleRateChanges = (event: Event): void => {
         const element: HTMLInputElement = <HTMLInputElement>event.target;
         const entryIndex: number = +(element.dataset.index!);
-        const value: number = clamp(8000, 96000 + 1, parseFloatWithDefault(element.value, 44100));
-        this._entries[entryIndex].sampleRate = value;
+        if (element.value === "") {
+            this._entries[entryIndex].sampleRate = 44100;
+        } else {
+            this._entries[entryIndex].sampleRate = +element.value;
+        }
     }
 
     private _whenRootKeyChanges = (event: Event): void => {
         const element: HTMLInputElement = <HTMLInputElement>event.target;
         const entryIndex: number = +(element.dataset.index!);
-        const value: number = parseFloatWithDefault(element.value, 60);
-        this._entries[entryIndex].rootKey = value;
+        if (element.value === "") {
+            this._entries[entryIndex].rootKey = 60;
+        } else {
+            this._entries[entryIndex].rootKey = +element.value;
+        }
         const rootKeyDisplay: HTMLSpanElement | null | undefined = element.parentNode?.parentNode?.querySelector(".add-sample-prompt-root-key-display");
         if (rootKeyDisplay != null) {
             const noteName: string = this._noteNameFromPitchNumber(this._entries[entryIndex].rootKey);
@@ -301,69 +292,31 @@ export class AddSamplesPrompt {
     private _whenChipWaveLoopStartChanges = (event: Event): void => {
         const element: HTMLInputElement = <HTMLInputElement>event.target;
         const entryIndex: number = +(element.dataset.index!);
-        const value: number | null = parseIntWithDefault(element.value, null);
-        this._entries[entryIndex].chipWaveLoopStart = value;
+        if (element.value === "") {
+            this._entries[entryIndex].chipWaveLoopStart = null;
+        } else {
+            this._entries[entryIndex].chipWaveLoopStart = +element.value;
+        }
     }
 
     private _whenChipWaveLoopEndChanges = (event: Event): void => {
         const element: HTMLInputElement = <HTMLInputElement>event.target;
         const entryIndex: number = +(element.dataset.index!);
-        const value: number | null = parseIntWithDefault(element.value, null);
-        this._entries[entryIndex].chipWaveLoopEnd = value;
+        if (element.value === "") {
+            this._entries[entryIndex].chipWaveLoopEnd = null;
+        } else {
+            this._entries[entryIndex].chipWaveLoopEnd = +element.value;
+        }
     }
 
     private _whenChipWaveStartOffsetChanges = (event: Event): void => {
         const element: HTMLInputElement = <HTMLInputElement>event.target;
         const entryIndex: number = +(element.dataset.index!);
-        const value: number | null = parseIntWithDefault(element.value, null);
-        this._entries[entryIndex].chipWaveStartOffset = value;
-    }
-
-    private _whenChipWaveLoopModeChanges = (event: Event): void => {
-        const element: HTMLSelectElement = <HTMLSelectElement>event.target;
-        const entryIndex: number = +(element.dataset.index!);
-        const newValue: number = +element.value;
-        if (newValue === -1) {
-            this._entries[entryIndex].chipWaveLoopMode = null;
+        if (element.value === "") {
+            this._entries[entryIndex].chipWaveStartOffset = null;
         } else {
-            this._entries[entryIndex].chipWaveLoopMode = newValue;
+            this._entries[entryIndex].chipWaveStartOffset = +element.value;
         }
-    }
-
-    private _whenChipWavePlayBackwardsChanges = (event: Event): void => {
-        const element: HTMLInputElement = <HTMLInputElement>event.target;
-        const entryIndex: number = +(element.dataset.index!);
-        const newValue: boolean = element.checked;
-        this._entries[entryIndex].chipWavePlayBackwards = newValue;
-    }
-
-    // @TODO: This is copy pasted from SongEditor, should probably be moved to
-    //        somewhere else that can be imported from both places.
-    private _copyTextToClipboard(text: string): void {
-        // Set as any to allow compilation without clipboard types (since, uh, I didn't write this bit and don't know the proper types library) -jummbus
-        let nav: any;
-        nav = navigator;
-
-        if (nav.clipboard && nav.clipboard.writeText) {
-            nav.clipboard.writeText(text).catch(() => {
-                window.prompt("Copy to clipboard:", text);
-            });
-            return;
-        }
-        const textField: HTMLTextAreaElement = document.createElement("textarea");
-        textField.textContent = text;
-        document.body.appendChild(textField);
-        textField.select();
-        const succeeded: boolean = document.execCommand("copy");
-        textField.remove();
-        this.container.focus({ preventScroll: true });
-        if (!succeeded) window.prompt("Copy this:", text);
-    }
-
-    private _whenCopyLinkPresetClicked = (event: Event): void => {
-        const element: HTMLButtonElement = <HTMLButtonElement>event.target;
-        const entryIndex: number = +(element.dataset.index!);
-        this._copyTextToClipboard(this._generateURLDataForEntry(this._entries[entryIndex]));
     }
 
     private _whenRemoveSampleClicked = (event: Event): void => {
@@ -417,16 +370,16 @@ export class AddSamplesPrompt {
         }
     }
 
-    private _parseURLs = (urls: string[], parseOldSyntax: boolean): ParsedEntries => {
+    private _parseURLs = (urls: string[]): ParsedEntries => {
         // @TODO: Duplicated code like this isn't great (in this case coming from Song.fromBase64String).
         function sliceForSampleRate(url: string): [string, number] {
             const newUrl = url.slice(0, url.indexOf(","));
-            const sampleRate = clamp(8000, 96000 + 1, parseFloatWithDefault(url.slice(url.indexOf(",") + 1), 44100));
+            const sampleRate = clamp(8000, 96000, parseFloat(url.slice(url.indexOf(",") + 1)));
             return [newUrl, sampleRate];
         }
         function sliceForRootKey(url: string): [string, number] {
             const newUrl = url.slice(0, url.indexOf("!"));
-            const rootKey = parseFloatWithDefault(url.slice(url.indexOf("!") + 1), 60);
+            const rootKey = parseFloat(url.slice(url.indexOf("!") + 1));
             return [newUrl, rootKey];
         }
         let useLegacySamples: boolean = false;
@@ -445,8 +398,6 @@ export class AddSamplesPrompt {
                         chipWaveLoopStart: null,
                         chipWaveLoopEnd: null,
                         chipWaveStartOffset: null,
-                        chipWaveLoopMode: null,
-                        chipWavePlayBackwards: false,
                     });
                 }
                 useLegacySamples = true;
@@ -460,8 +411,6 @@ export class AddSamplesPrompt {
                         chipWaveLoopStart: null,
                         chipWaveLoopEnd: null,
                         chipWaveStartOffset: null,
-                        chipWaveLoopMode: null,
-                        chipWavePlayBackwards: false,
                     });
                 }
                 useNintariboxSamples = true;
@@ -475,8 +424,6 @@ export class AddSamplesPrompt {
                         chipWaveLoopStart: null,
                         chipWaveLoopEnd: null,
                         chipWaveStartOffset: null,
-                        chipWaveLoopMode: null,
-                        chipWavePlayBackwards: false,
                     });
                 }
                 useMarioPaintboxSamples = true;
@@ -488,8 +435,6 @@ export class AddSamplesPrompt {
                 let chipWaveLoopStart: number | null = null;
                 let chipWaveLoopEnd: number | null = null;
                 let chipWaveStartOffset: number | null = null;
-                let chipWaveLoopMode: number | null = null;
-                let chipWavePlayBackwards: boolean = false;
                 let optionsStartIndex: number = url.indexOf("!");
                 let optionsEndIndex: number = -1;
                 let parsedSampleOptions: boolean = false;
@@ -501,56 +446,44 @@ export class AddSamplesPrompt {
                             const optionCode: string = rawOption.charAt(0);
                             const optionData: string = rawOption.slice(1, rawOption.length);
                             if (optionCode === "s") {
-                                sampleRate = clamp(8000, 96000 + 1, parseFloatWithDefault(optionData, 44100));
+                                sampleRate = clamp(8000, 96000, parseFloat(optionData));
                             } else if (optionCode === "r") {
-                                rootKey = parseFloatWithDefault(optionData, 60);
+                                rootKey = parseFloat(optionData);
                             } else if (optionCode === "p") {
                                 percussion = true;
                             } else if (optionCode === "a") {
-                                chipWaveLoopStart = parseIntWithDefault(optionData, null);
+                                chipWaveLoopStart = parseInt(optionData);
                             } else if (optionCode === "b") {
-                                chipWaveLoopEnd = parseIntWithDefault(optionData, null);
+                                chipWaveLoopEnd = parseInt(optionData);
                             } else if (optionCode === "c") {
-                                chipWaveStartOffset = parseIntWithDefault(optionData, null);
-                            } else if (optionCode === "d") {
-                                chipWaveLoopMode = parseIntWithDefault(optionData, null);
-                                if (chipWaveLoopMode != null) {
-                                    // @TODO: Error-prone. This should be
-                                    // automatically derived from the list of
-                                    // available loop modes.
-                                    chipWaveLoopMode = clamp(0, 3 + 1, chipWaveLoopMode);
-                                }
-                            } else if (optionCode === "e") {
-                                chipWavePlayBackwards = true;
+                                chipWaveStartOffset = parseInt(optionData);
                             }
                         }
                         urlSliced = url.slice(optionsEndIndex + 1, url.length);
                         parsedSampleOptions = true;
                     }
                 }
-                if (parseOldSyntax) {
-                    if (!parsedSampleOptions) {
-                        if (url.indexOf("@") != -1) {
-                            urlSliced = url.replaceAll("@", "");
-                            percussion = true;
-                        }
-                        if (url.indexOf(",") != -1 && url.indexOf("!") != -1) {
-                            if (url.indexOf(",") < url.indexOf("!")) {
-                                [urlSliced, rootKey] = sliceForRootKey(urlSliced);
-                                [urlSliced, sampleRate] = sliceForSampleRate(urlSliced);
-                            }
-                            else {
-                                [urlSliced, sampleRate] = sliceForSampleRate(urlSliced);
-                                [urlSliced, rootKey] = sliceForRootKey(urlSliced);
-                            }
+                if (!parsedSampleOptions) {
+                    if (url.indexOf("@") != -1) {
+                        urlSliced = url.replaceAll("@", "");
+                        percussion = true;
+                    }
+                    if (url.indexOf(",") != -1 && url.indexOf("!") != -1) {
+                        if (url.indexOf(",") < url.indexOf("!")) {
+                            [urlSliced, rootKey] = sliceForRootKey(urlSliced);
+                            [urlSliced, sampleRate] = sliceForSampleRate(urlSliced);
                         }
                         else {
-                            if (url.indexOf(",") != -1) {
-                                [urlSliced, sampleRate] = sliceForSampleRate(urlSliced);
-                            }
-                            if (url.indexOf("!") != -1) {
-                                [urlSliced, rootKey] = sliceForRootKey(urlSliced);
-                            }
+                            [urlSliced, sampleRate] = sliceForSampleRate(urlSliced);
+                            [urlSliced, rootKey] = sliceForRootKey(urlSliced);
+                        }
+                    }
+                    else {
+                        if (url.indexOf(",") != -1) {
+                            [urlSliced, sampleRate] = sliceForSampleRate(urlSliced);
+                        }
+                        if (url.indexOf("!") != -1) {
+                            [urlSliced, rootKey] = sliceForRootKey(urlSliced);
                         }
                     }
                 }
@@ -562,52 +495,41 @@ export class AddSamplesPrompt {
                     chipWaveLoopStart: chipWaveLoopStart,
                     chipWaveLoopEnd: chipWaveLoopEnd,
                     chipWaveStartOffset: chipWaveStartOffset,
-                    chipWaveLoopMode: chipWaveLoopMode,
-                    chipWavePlayBackwards: chipWavePlayBackwards,
                 });
             }
         }
         return { entries: parsedEntries };
     }
 
-    private _generateURLDataForEntry = (entry: SampleEntry): string => {
-        const url: string = entry.url.trim();
-        const sampleRate: number = entry.sampleRate;
-        const rootKey: number = entry.rootKey;
-        const percussion: boolean = entry.percussion;
-        const chipWaveLoopStart: number | null = entry.chipWaveLoopStart;
-        const chipWaveLoopEnd: number | null = entry.chipWaveLoopEnd;
-        const chipWaveStartOffset: number | null = entry.chipWaveStartOffset;
-        const chipWaveLoopMode: number | null = entry.chipWaveLoopMode;
-        const chipWavePlayBackwards: boolean = entry.chipWavePlayBackwards;
-        const urlInLowerCase: string = url.toLowerCase();
-        const isBundledSamplePack: boolean = (
-            urlInLowerCase === "legacysamples"
-            || urlInLowerCase === "nintariboxsamples"
-            || urlInLowerCase === "mariopaintboxsamples"
-        );
-        const options: string[] = [];
-        if (sampleRate !== 44100) options.push("s" + sampleRate);
-        if (rootKey !== 60) options.push("r" + rootKey);
-        if (percussion) options.push("p");
-        if (chipWaveLoopStart != null) options.push("a" + chipWaveLoopStart);
-        if (chipWaveLoopEnd != null) options.push("b" + chipWaveLoopEnd);
-        if (chipWaveStartOffset != null) options.push("c" + chipWaveStartOffset);
-        if (chipWaveLoopMode != null) options.push("d" + chipWaveLoopMode);
-        if (chipWavePlayBackwards) options.push("e");
-        if (isBundledSamplePack || options.length <= 0) {
-            return url;
-        } else {
-            return "!" + options.join(",") + "!" + url;
-        }
-    }
-
     private _generateURLData = (): string => {
         let output = "";
         for (const entry of this._entries) {
             const url: string = entry.url.trim();
+            const sampleRate: number = entry.sampleRate;
+            const rootKey: number = entry.rootKey;
+            const percussion: boolean = entry.percussion;
+            const chipWaveLoopStart: number | null = entry.chipWaveLoopStart;
+            const chipWaveLoopEnd: number | null = entry.chipWaveLoopEnd;
+            const chipWaveStartOffset: number | null = entry.chipWaveStartOffset;
             if (url === "") continue;
-            output += "|" + this._generateURLDataForEntry(entry);
+            const urlInLowerCase: string = url.toLowerCase();
+            const isBundledSamplePack: boolean = (
+                urlInLowerCase === "legacysamples"
+                || urlInLowerCase === "nintariboxsamples"
+                || urlInLowerCase === "mariopaintboxsamples"
+            );
+            const options: string[] = [];
+            if (sampleRate !== 44100) options.push("s" + sampleRate);
+            if (rootKey !== 60) options.push("r" + rootKey);
+            if (percussion) options.push("p");
+            if (chipWaveLoopStart != null) options.push("a" + chipWaveLoopStart);
+            if (chipWaveLoopEnd != null) options.push("b" + chipWaveLoopEnd);
+            if (chipWaveStartOffset != null) options.push("c" + chipWaveStartOffset);
+            if (isBundledSamplePack || options.length <= 0) {
+                output += "|" + url;
+            } else {
+                output += "|!" + options.join(",") + "!" + url;
+            }
         }
         return output;
     }
@@ -658,65 +580,44 @@ export class AddSamplesPrompt {
             const optionsVisible: boolean = Boolean(this._entryOptionsDisplayStates[entryIndex]);
             const urlInput: HTMLInputElement = input({ style: "flex-grow: 1; margin-left: 1em; width: 100%;", value: entry.url });
             const sampleRateStepper: HTMLInputElement = input({ style: "flex-grow: 1; margin-left: 1em; width: 100%;", type: "number", value: "" + entry.sampleRate, min: "8000", max: "96000", step: "1" });
-            const rootKeyStepper: HTMLInputElement = input({ style: "flex-grow: 1; margin-left: 1em; width: 100%;", type: "number", value: "" + entry.rootKey, min: "0", max: Config.maxPitch + Config.pitchesPerOctave, step: "1" });
+            const rootKeyStepper: HTMLInputElement = input({ style: "flex-grow: 1; margin-left: 1em; width: 100%;", type: "number", value: "" + entry.rootKey, min: "0", max: Config.maxPitch, step: "1" });
             const rootKeyDisplay: HTMLSpanElement = span({ class: "add-sample-prompt-root-key-display", style: "margin-left: 0.4em; width: 3em; text-align: left; text-overflow: ellipsis; overflow: hidden; flex-shrink: 0;" }, `(${this._noteNameFromPitchNumber(entry.rootKey)})`);
             const percussionBox: HTMLInputElement = input({ style: "width: 1em; margin-left: 1em;", type: "checkbox" });
             const chipWaveLoopStartStepper: HTMLInputElement = input({ style: "flex-grow: 1; margin-left: 1em; width: 100%;", type: "number", value: "" + (entry.chipWaveLoopStart != null ? entry.chipWaveLoopStart : ""), min: "0", step: "1" });
             const chipWaveLoopEndStepper: HTMLInputElement = input({ style: "flex-grow: 1; margin-left: 1em; width: 100%;", type: "number", value: "" + (entry.chipWaveLoopEnd != null ? entry.chipWaveLoopEnd : ""), min: "0", step: "1" });
             const chipWaveStartOffsetStepper: HTMLInputElement = input({ style: "flex-grow: 1; margin-left: 1em; width: 100%;", type: "number", value: "" + (entry.chipWaveStartOffset != null ? entry.chipWaveStartOffset : ""), min: "0", step: "1" });
-            const chipWaveLoopModeSelect: HTMLSelectElement = select({ style: "width: 100%; flex-grow: 1; margin-left: 0.5em;" },
-                option({ value: -1 }, ""),
-                option({ value: 0 }, "Loop"),
-                option({ value: 1 }, "Ping-Pong"),
-                option({ value: 2 }, "Play Once"),
-                option({ value: 3 }, "Play Loop Once"),
-            );
-            if (entry.chipWaveLoopMode != null) {
-                chipWaveLoopModeSelect.value = "" + entry.chipWaveLoopMode;
-            }
-            const chipWavePlayBackwardsBox: HTMLInputElement = input({ type: "checkbox", style: "width: 1em; padding: 0; margin-left: auto; margin-right: auto;" });
-            chipWavePlayBackwardsBox.checked = entry.chipWavePlayBackwards;
             const sampleName: string = this._getSampleName(entry);
             percussionBox.checked = entry.percussion;
-            const copyLinkPresetButton: HTMLButtonElement = button({ style: "height: auto; min-height: var(--button-size);", title: "For use with \"Add multiple samples\"" }, "Copy link preset");
-            const removeButton: HTMLButtonElement = button({ style: "height: auto; min-height: var(--button-size); margin-left: 0.5em;" }, "Remove");
+            const removeButton: HTMLButtonElement = button({ style: "height: auto; min-height: var(--button-size);" }, "Remove");
             const moveUpButton: HTMLButtonElement = button({ style: "height: auto; min-height: var(--button-size); margin-left: 0.5em;" }, SVG.svg({ width: "16", height: "16", viewBox: "-13 -14 26 26", "pointer-events": "none", style: "width: 100%; height: 100%;" }, SVG.path({ d: "M -6 6 L 0 -6 L 6 6 z", fill: ColorConfig.primaryText })));
             const moveDownButton: HTMLButtonElement = button({ style: "height: auto; min-height: var(--button-size); margin-left: 0.5em;" }, SVG.svg({ width: "16", height: "16", viewBox: "-13 -14 26 26", "pointer-events": "none", style: "width: 100%; height: 100%;" }, SVG.path({ d: "M -6 -6 L 6 -6 L 0 6 z", fill: ColorConfig.primaryText })));
             const optionsContainer: HTMLDetailsElement = details(
-                { open: optionsVisible, style: "margin-bottom: 2em; margin-top: 1em;" },
-                summary({ style: "margin-bottom: 1em;" }, "Options"),
+                { open: optionsVisible },
+                summary({ style: "margin-bottom: 0.5em;" }, "Options"),
                 div({ style: "display: flex; flex-direction: row; align-items: center; justify-content: flex-end; margin-bottom: 0.5em;" },
-                    div({ style: `flex-shrink: 0; :text-align: right; color: ${ColorConfig.primaryText};` }, span({ title: "What rate to resample to" }, "Sample rate")),
+                    div({ style: `flex-shrink: 0; :text-align: right; color: ${ColorConfig.primaryText};` }, "Sample rate"),
                     sampleRateStepper
                 ),
                 div({ style: "display: flex; flex-direction: row; align-items: center; justify-content: flex-end; margin-bottom: 0.5em;" },
-                    div({ style: `text-align: right; color: ${ColorConfig.primaryText}; flex-shrink: 0;` }, span({ title: "Pitch where the sample is played as-is" }, "Root key")),
+                    div({ style: `text-align: right; color: ${ColorConfig.primaryText}; flex-shrink: 0;` }, "Root key"),
                     rootKeyDisplay,
                     rootKeyStepper
                 ),
-                div({ style: "display: flex; flex-direction: row; align-items: center; justify-content: space-between; margin-bottom: 0.5em;" },
+                div({ style: "display: flex; flex-direction: row; align-items: center; justify-content: space-around; margin-bottom: 0.5em;" },
                     div({ style: `text-align: right; color: ${ColorConfig.primaryText};` }, "Percussion (pitch doesn't change with key)"),
                     percussionBox
                 ),
                 div({ style: "display: flex; flex-direction: row; align-items: center; justify-content: flex-end; margin-bottom: 0.5em;" },
-                    div({ style: `flex-shrink: 0; text-align: right; color: ${ColorConfig.primaryText};` }, span({ title: "Applies to the \"Loop Start\" loop control option of the preset created for this sample" }, "Loop Start")),
+                    div({ style: `flex-shrink: 0; text-align: right; color: ${ColorConfig.primaryText};` }, "Loop Start"),
                     chipWaveLoopStartStepper
                 ),
                 div({ style: "display: flex; flex-direction: row; align-items: center; justify-content: flex-end; margin-bottom: 0.5em;" },
-                    div({ style: `flex-shrink: 0; text-align: right; color: ${ColorConfig.primaryText};` }, span({ title: "Applies to the \"Loop End\" loop control option of the preset created for this sample" }, "Loop End")),
+                    div({ style: `flex-shrink: 0; text-align: right; color: ${ColorConfig.primaryText};` }, "Loop End"),
                     chipWaveLoopEndStepper
                 ),
                 div({ style: "display: flex; flex-direction: row; align-items: center; justify-content: flex-end; margin-bottom: 0.5em;" },
-                    div({ style: `flex-shrink: 0; text-align: right; color: ${ColorConfig.primaryText};` }, span({ title: "Applies to the \"Offset\" loop control option of the preset created for this sample" }, "Sample Start Offset")),
+                    div({ style: `flex-shrink: 0; text-align: right; color: ${ColorConfig.primaryText};` }, "Sample Start Offset"),
                     chipWaveStartOffsetStepper
-                ),
-                div({ style: "display: flex; flex-direction: row; align-items: center; justify-content: flex-end; margin-bottom: 0.5em;" },
-                    div({ style: `flex-shrink: 0; text-align: right; color: ${ColorConfig.primaryText};` }, span({ title: "Applies to the \"Loop Mode\" loop control option of the preset created for this sample" }, "Loop Mode")),
-                    chipWaveLoopModeSelect
-                ),
-                div({ style: "display: flex; flex-direction: row; align-items: center; justify-content: flex-end; margin-bottom: 0.5em;" },
-                    div({ style: `flex-shrink: 0; text-align: right; color: ${ColorConfig.primaryText};` }, span({ title: "Applies to the \"Backwards\" loop control option of the preset created for this sample" }, "Backwards")),
-                    chipWavePlayBackwardsBox
                 ),
             );
             urlInput.dataset.index = "" + entryIndex;
@@ -726,14 +627,11 @@ export class AddSamplesPrompt {
             chipWaveLoopStartStepper.dataset.index = "" + entryIndex;
             chipWaveLoopEndStepper.dataset.index = "" + entryIndex;
             chipWaveStartOffsetStepper.dataset.index = "" + entryIndex;
-            chipWaveLoopModeSelect.dataset.index = "" + entryIndex;
-            chipWavePlayBackwardsBox.dataset.index = "" + entryIndex;
-            copyLinkPresetButton.dataset.index = "" + entryIndex;
             removeButton.dataset.index = "" + entryIndex;
             moveUpButton.dataset.index = "" + entryIndex;
             moveDownButton.dataset.index = "" + entryIndex;
             optionsContainer.dataset.index = "" + entryIndex;
-            const bottomButtons: HTMLDivElement = div({ style: "display: flex; flex-direction: row; align-items: center; justify-content: flex-end;" }, copyLinkPresetButton, removeButton);
+            const bottomButtons: HTMLDivElement = div({ style: "display: flex; flex-direction: row; align-items: center; justify-content: flex-end;" }, removeButton);
             if (canMoveUp) {
                 bottomButtons.appendChild(moveUpButton);
             }
@@ -763,9 +661,6 @@ export class AddSamplesPrompt {
             chipWaveLoopStartStepper.addEventListener("change", this._whenChipWaveLoopStartChanges);
             chipWaveLoopEndStepper.addEventListener("change", this._whenChipWaveLoopEndChanges);
             chipWaveStartOffsetStepper.addEventListener("change", this._whenChipWaveStartOffsetChanges);
-            chipWaveLoopModeSelect.addEventListener("change", this._whenChipWaveLoopModeChanges);
-            chipWavePlayBackwardsBox.addEventListener("change", this._whenChipWavePlayBackwardsChanges);
-            copyLinkPresetButton.addEventListener("click", this._whenCopyLinkPresetClicked);
             removeButton.addEventListener("click", this._whenRemoveSampleClicked);
             if (canMoveUp) {
                 moveUpButton.addEventListener("click", this._whenMoveSampleUpClicked);
